wip: rename adapter
