+++ src/kairos/workspace/adapter/mongodb/insert-workspace.test.ts
@@ -1,6 +1,6 @@
 import { workspacesCollection } from '@/framework/mongodb/collection/workspaces-collection';
 import type { WorkspacesCollection } from '@/framework/mongodb/schema/workspaces-collection-schema';
-import { makeInsertWorkspace } from '@/kairos/workspace/adapter/repository/insert-workspace';
+import { makeInsertWorkspace } from '@/kairos/workspace/adapter/mongodb/insert-workspace';
 import { workspaceSlugConflictError } from '@/kairos/workspace/errors';
 import { expectAsyncToThrow } from '@tests/__vitest__/expect-async-to-throw';
 import { integrationTest } from '@tests/__vitest__/integration-test';
