+++ src/kairos/workspace/create-workspace/create-workspace.compose.ts
@@ -1,5 +1,5 @@
 import { workspacesCollection } from '@/framework/mongodb/collection/workspaces-collection';
-import { makeInsertWorkspace } from '@/kairos/workspace/adapter/repository/insert-workspace';
+import { makeInsertWorkspace } from '@/kairos/workspace/adapter/mongodb/insert-workspace';
 import { makeCreateWorkspace } from '@/kairos/workspace/create-workspace/create-workspace';
 import type { InsertWorkspace } from '@/kairos/workspace/create-workspace/insert-workspace.type';
